Split up build utils
