+++ tools/gulp/packaging/build-tasks-gulp.ts
@@ -3,7 +3,8 @@ import {join} from 'path';
 import {main as tsc} from '@angular/tsc-wrapped';
 import {SOURCE_ROOT, DIST_ROOT} from '../constants';
 import {sequenceTask, sassBuildTask, copyTask, triggerLivereload} from '../util/task_helpers';
-import {buildPackageBundles, composeRelease} from './build-functions';
+import {composeRelease} from './build-release';
+import {buildPackageBundles} from './build-bundles';
 
 // There are no type definitions available for these imports.
 const inlineResources = require('../../../scripts/release/inline-resources');
