+++ tools/gulp/packaging/metadata-inlining.ts
