+++ tools/gulp/packaging/sourcemap-remap.ts
@@ -1,0 +1,13 @@
+// There are no type definitions available for these imports.
+const sorcery = require('sorcery');
+
+/**
+ * Finds the original sourcemap of the file and maps it to the current file.
+ * This is useful when multiple transformation happen (e.g TSC -> Rollup -> Uglify)
+ **/
+export async function remapSourcemap(sourceFile: string) {
+  // Once sorcery loaded the chain of sourcemaps, the new sourcemap will be written asynchronously.
+  return (await sorcery.load(sourceFile)).write();
+}
+
+
